split SpanningInfo from DetailStatusView
